add chroma
